Implementation/rag_phase2/requirements.txt
sentence-transformers
faiss-cpu
psycopg2-binary
pdfplumber
numpy
